Unify package names and groupId to io.packt
