+++ Chapter05-metrics/pom.xml
@@ -2,20 +2,17 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
   <modelVersion>4.0.0</modelVersion>
-  <groupId>com.packt.microprofile-book</groupId>
-  <artifactId>metrics</artifactId>
-  <version>1.0-SNAPSHOT</version>
-
-  <packaging>war</packaging>
-
   <parent>
-    <groupId>io.microprofile</groupId>
+    <groupId>io.packt</groupId>
     <artifactId>handson-microprofile</artifactId>
     <version>1.0-SNAPSHOT</version>
     <relativePath>../pom.xml</relativePath>
   </parent>
+
+  <groupId>io.packt.metrics</groupId>
+  <artifactId>metrics</artifactId>
+  <packaging>war</packaging>
 
   <properties>
     <maven.compiler.source>1.8</maven.compiler.source>
