+++ Chapter05-opentracing/pom.xml
@@ -3,13 +3,13 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>io.microprofile</groupId>
+        <groupId>io.packt</groupId>
         <artifactId>handson-microprofile</artifactId>
         <relativePath>../pom.xml</relativePath>
         <version>1.0-SNAPSHOT</version>
     </parent>
 
-    <groupId>io.microprofile.opentracing</groupId>
+    <groupId>io.packt.ot</groupId>
     <artifactId>opentracing</artifactId>
     <name>OpenTracing</name>
     <packaging>war</packaging>
