+++ Chapter05-opentracing/src/main/java/io/packt/ot/rest/GreetingService.java
@@ -1,4 +1,4 @@
-package io.pckt.ot.rest;
+package io.packt.ot.rest;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
