+++ Chapter05-opentracing/src/main/resources/META-INF/microprofile-config.properties
@@ -1,1 +1,1 @@
-io.pckt.ot.rest.GreetingService/mp-rest/url=http://localhost:8080
+io.packt.ot.rest.GreetingService/mp-rest/url=http://localhost:8080
