+++ Chapter06-restclient/pom.xml
@@ -3,13 +3,13 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>io.microprofile</groupId>
+        <groupId>io.packt</groupId>
         <artifactId>handson-microprofile</artifactId>
         <relativePath>../pom.xml</relativePath>
         <version>1.0-SNAPSHOT</version>
     </parent>
 
-    <groupId>io.microprofile.jwt</groupId>
+    <groupId>io.packt.restc</groupId>
     <artifactId>rest-client</artifactId>
     <name>REST Client</name>
     <packaging>war</packaging>
