+++ Chapter06-restclient/src/main/java/io/packt/restc/contract/WorldClockApi.java
@@ -1,6 +1,4 @@
-package io.pckt.restc.contract;
-
-import java.util.concurrent.CompletionStage;
+package io.packt.restc.contract;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
@@ -8,20 +6,21 @@ import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 
+import org.eclipse.microprofile.rest.client.annotation.ClientHeaderParam;
 import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
 
 @Path("/api/json")
-@RegisterRestClient(baseUri = WorldClockApiWithHeaders.BASE_URL)
-public interface WorldClockApiAsync {
-    String BASE_URL = "http://worldclockapi.com/api/json";
+@RegisterRestClient(baseUri = WorldClockApi.BASE_URL)
+public interface WorldClockApi {
+    static final String BASE_URL = "http://worldclockapi.com/api/json";
 
     @GET
     @Path("/utc/now")
     @Produces(MediaType.APPLICATION_JSON)
-    CompletionStage<Now> utc();
+    Now utc();
 
     @GET
     @Path("{tz}/now")
     @Produces(MediaType.APPLICATION_JSON)
-    CompletionStage<Now> tz(@PathParam("tz") String tz);
+    Now tz(@PathParam("tz") String tz);
 }
