+++ Chapter06-restclient/src/main/java/io/packt/restc/contract/WorldClockApiAsync.java
@@ -1,4 +1,6 @@
-package io.pckt.restc.contract;
+package io.packt.restc.contract;
+
+import java.util.concurrent.CompletionStage;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
@@ -6,21 +8,20 @@ import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 
-import org.eclipse.microprofile.rest.client.annotation.ClientHeaderParam;
 import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
 
 @Path("/api/json")
-@RegisterRestClient(baseUri = WorldClockApi.BASE_URL)
-public interface WorldClockApi {
-    static final String BASE_URL = "http://worldclockapi.com/api/json";
+@RegisterRestClient(baseUri = WorldClockApiWithHeaders.BASE_URL)
+public interface WorldClockApiAsync {
+    String BASE_URL = "http://worldclockapi.com/api/json";
 
     @GET
     @Path("/utc/now")
     @Produces(MediaType.APPLICATION_JSON)
-    Now utc();
+    CompletionStage<Now> utc();
 
     @GET
     @Path("{tz}/now")
     @Produces(MediaType.APPLICATION_JSON)
-    Now tz(@PathParam("tz") String tz);
+    CompletionStage<Now> tz(@PathParam("tz") String tz);
 }
