+++ Chapter06-restclient/src/main/java/io/packt/restc/contract/WorldClockApiWithHeaders.java
@@ -1,4 +1,4 @@
-package io.pckt.restc.contract;
+package io.packt.restc.contract;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
