+++ Chapter06-restclient/src/main/java/io/packt/restc/services/WorldClockCDIUser.java
@@ -1,5 +1,7 @@
 package io.packt.restc.services;
 
+import io.packt.restc.contract.Now;
+import io.packt.restc.contract.WorldClockApi;
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 import javax.ws.rs.GET;
@@ -7,8 +9,6 @@ import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 
-import io.pckt.restc.contract.Now;
-import io.pckt.restc.contract.WorldClockApi;
 import org.eclipse.microprofile.rest.client.inject.RestClient;
 
 @ApplicationScoped
