+++ pom.xml
@@ -3,7 +3,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>io.microprofile</groupId>
+    <groupId>io.packt</groupId>
     <artifactId>handson-microprofile</artifactId>
     <name>Hands-On-Enterprise-Java-Microservices-with-Eclipse-MicroProfile - Parent pom</name>
     <version>1.0-SNAPSHOT</version>
